backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
requests
python-multipart
streamlit
numpy
tqdm
google-generativeai
sentence-transformers
scikit-learn
